+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- **queue** stat instead of **out_queue** because **out_queue** now shows diff between outputed and transformed metrics.
+- Send **groxy** state as metrics to Graphite.
+
 ## [0.7.0] - 2021-01-24
 ### Changed
 - Logging with **logrus** library.